requirements.txt
uuid~=1.30
aiofiles~=0.6.0
Markdown~=3.3.6
python-magic~=0.4.25
Pillow~=9.0.1
python-olm~=3.1.3
matrix-nio[e2e]~=0.19.0
datetime